BZ64006: Provide default configuration source that "works"
Although this will likely be revisited in Tomcat 10, the plan of the
configuration source was not to break existing Tomcat 9 code. As a
result, use a default source based on the user folder when none is
specified. Although it would be better to log this, extra logs like this
are usually not well received so skip that as well for now.
